Refine validate command structure and load context
