+++ src/cli/mod.rs
@@ -1,0 +1,22 @@
+mod app;
+mod validate;
+
+use std::process::ExitCode;
+
+use clap::Parser;
+
+use app::{Cli, Commands};
+
+pub fn run() -> ExitCode {
+    let cli = match Cli::try_parse() {
+        Ok(cli) => cli,
+        Err(error) => {
+            let _ = error.print();
+            return ExitCode::from(error.exit_code() as u8);
+        }
+    };
+
+    match cli.command {
+        Commands::Validate(args) => validate::run(args),
+    }
+}
